requirement.txt
allure-pytest==2.8.13
allure-python-commons==2.8.13
attrs==19.3.0
beeprint==2.4.10
certifi==2020.4.5.1
chardet==3.0.4
enum34==1.1.10
graphql-core==3.1.0
idna==2.9
importlib-metadata==1.6.0
jsonpath==0.82
lxml==4.5.0
more-itertools==8.2.0
namedlist==1.7
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
pytest-rerunfailures==9.0
PyYAML==5.3.1
requests==2.23.0
sgqlc==10.1
shell==1.0.1
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.8
urwid==2.1.0
wcwidth==0.1.9
zipp==3.1.0
